First version of fixture bot, deleting comments still not working
